projects/GraphQLSample/graphql/requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
Flask~=2.3.3
SQLAlchemy~=2.0.25
ariadne~=0.23.0